Special case tensor argument in kernel signature for pyop2 backend
The tensor argument in the kernel signature is a fixed-shape double
array with the shape of the local tensor. This does not work for the
case of a rank 0 tensor i.e. a scalar, since the tensor shape in this
case is empty. We use the shape (1,) instead in this case.