Pass cell orientations for all domain types
It is not just cell integrals that need to know about cell orientations,
but also facet integrals.  To simplify handling in Firedrake, switch to
reading the cell orientations as if they were another
coefficient (rather than a special-cased single-indirection field).
This means we don't have to do any special different magic for interior
facets.