Don't use function or argument cache if in PyOP2 mode
If we're in PyOP2 mode, we always use the optimised quadrature
transformer (even if optimisation is off).  As a result, we would splat
the symbol on the operand in the optimised math_function visitor, but
this led to incorrect behaviour when compiling forms where the same
symbol appeared in more than one context.  For example:

   compile_form((tanh(f) + sin(f) + cos(f))*dx)

would generate code as if it had compiled 3*tanh(sin(cos(f)))*dx.  Fix
this, by always generating a new operand in the base quadrature
transformer for these cases.